requirements.txt

certifi
dnspython==2.4.2
faiss-cpu
ipython==8.18.1
jsonpatch==1.33
streamlit-lottie
langchain
langchain_community
langchain-experimental
langchainhub
langchain-text-splitters
langdetect
llama-index-cli
llama-index-core
llama-index-embeddings-langchain
Markdown==3.6
markdown-it-py==3.0.0
numpy==1.26.4
openai==0.28.0
opencv-python
openpyxl
pandas==2.2.2
pypdf2
python-dotenv
PyMuPDF
rank_bm25==0.2.2
requests
sentence-transformers
safetensors==0.4.4
scikit-learn==1.0.2
scikit-optimize==0.10.2
scipy
semantic-version==2.10.0
streamlit==1.31.0
streamlit_pdf_viewer==0.0.16
tensorflow==2.16.2
tf-keras
tokenizer==3.4.3
tokenizers==0.19.1
transformers==4.43.4
unstructured==0.12.5
unstructured-client==0.18.0
unstructured-inference==0.7.23
